backend/pyproject.toml
[project]
name = "undriven"
version = "0.2.0"
description = "Track car miles replaced by non-car trips"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.28",
]

[tool.pytest.ini_options]
pythonpath = [".."]
